# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: eSPX Management JSON API
  version: 1.0.0
  description: |
    Machine-readable contract for `/api/v1` JSON routes served by `cmd/management`
    (`internal/management` + `internal/adminapi`). HTMX `/admin/*` HTML routes are excluded.
    HTTP 501 stub routes (deferred dashboards/reports) are excluded — see `internal/adminapi/stub_routes.go` and `internal/openapi/stubs.go`.
servers:
  - url: http://127.0.0.1:8188
    description: Local management (default SERVER_PORT)
components:
  securitySchemes:
    AdminAPIKey:
      type: apiKey
      in: header
      name: X-Admin-API-Key
      description: Shared admin API key (`ADMIN_API_KEY`).
    SessionCookie:
      type: apiKey
      in: cookie
      name: accessToken
      description: JWT session cookie set by `POST /api/v1/auth/login`.
    ConsentHMAC:
      type: apiKey
      in: header
      name: X-Consent-Signature
      description: HMAC-SHA256 of raw request body (`CONSENT_HMAC_SECRET`).
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    GenericObject:
      type: object
      additionalProperties: true
...
    DashboardPayload:
      type: object
      additionalProperties: true
...
paths:
...
  /api/v1/cost-sync/credentials:
    get:
      tags: [Integrations]
      summary: GET /api/v1/cost-sync/credentials
      operationId: get_cost_sync_credentials
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/cost-sync/credentials/{network}:
    delete:
      tags: [Integrations]
      summary: DELETE /api/v1/cost-sync/credentials/{network}
      operationId: delete_cost_sync_credentials_network
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Integrations]
      summary: PUT /api/v1/cost-sync/credentials/{network}
      operationId: put_cost_sync_credentials_network
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/cost-sync/history:
    get:
      tags: [Integrations]
      summary: GET /api/v1/cost-sync/history
      operationId: get_cost_sync_history
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/cost-sync/run:
    post:
      tags: [Integrations]
      summary: POST /api/v1/cost-sync/run
      operationId: post_cost_sync_run
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/dashboards/campaign/{id}:
    get:
      tags: [Reporting]
      summary: GET /api/v1/dashboards/campaign/{id}
      operationId: get_dashboards_campaign_id
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardPayload'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/dashboards/operator:
    get:
      tags: [Reporting]
      summary: GET /api/v1/dashboards/operator
      operationId: get_dashboards_operator
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardPayload'
        '401':
          $ref: '#/components/responses/Unauthorized'
...